+++ src/nvmatrixlib/matrix.py
@@ -6,22 +6,34 @@ License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 import tkinter as tk
 from nvmatrixlib.nvmatrix_globals import *
+from nvmatrixlib.node import Node
 
 
 class Matrix:
     """Represent a matrix of relationships. 
     
+    Public methods:
+        set_nodes -- Loop through all nodes, setting states.
+        get_nodes -- Loop through all nodes, modifying the scenes according to the states.
+    
+    The visual part consists of one frame per column, each containing 
+    one node per row. 
+    The logical part consists of one dictionary per element type (protected instance variables):
+    {scene ID: {element Id: node}}
     """
 
     def __init__(self, master, novel):
-        """Initialize the instance variables.
+        """Draw the matrix with blank nodes.
         
         Positional arguments:
             novel -- Novel: Project reference.
+            
         """
         self._novel = novel
-        colorsFalse = (('white', 'gray95'), ('gray85', 'gray80'))
-        colorsTrue = (('blue4', 'red4'), ('blue3', 'red3'))
+        colorsBackground = (('white', 'gray95'), ('gray85', 'gray80'))
+        colorsCharacter = (('goldenrod3', 'orange3'), ('goldenrod4', 'orange4'))
+        colorsLocation = (('brown3', 'red3'), ('brown4', 'red4'))
+        colorsItem = (('green3', 'blue3'), ('green4', 'blue4'))
         row = 0
         bgr = row % 2
         col = 0
@@ -29,7 +41,7 @@ class Matrix:
         columns = []
         columns.append(tk.Frame(master))
         columns[col].pack(side=tk.LEFT)
-        tk.Label(columns[col], text=' ', bg=colorsFalse[bgr][bgc]).pack(fill=tk.X)
+        tk.Label(columns[col], text=' ', bg=colorsBackground[bgr][bgc]).pack(fill=tk.X)
 
         #--- Scene title column.
         for chId in novel.chapters:
@@ -38,18 +50,16 @@ class Matrix:
                 bgr = row % 2
                 tk.Label(columns[col],
                          text=novel.scenes[scId].title,
-                         bg=colorsFalse[bgr][bgc],
+                         bg=colorsBackground[bgr][bgc],
                          justify=tk.LEFT,
                          anchor=tk.W
                          ).pack(fill=tk.X)
 
-        #--- Initialize the matrix data.
-        self.sceneCharacters = {}
+        #--- Character columns.
+        self._characterNodes = {}
         for chId in novel.chapters:
             for scId in novel.chapters[chId].srtScenes:
-                self.sceneCharacters[scId] = {}
-
-        #--- Character columns.
+                self._characterNodes[scId] = {}
         for crId in novel.characters:
             row = 0
             bgr = row % 2
@@ -59,7 +69,7 @@ class Matrix:
             columns[col].pack(side=tk.LEFT)
             tk.Label(columns[col],
                      text=novel.characters[crId].title,
-                     bg=colorsFalse[bgr][bgc],
+                     bg=colorsBackground[bgr][bgc],
                      justify=tk.LEFT,
                      anchor=tk.W
                      ).pack(fill=tk.X)
@@ -68,61 +78,125 @@ class Matrix:
                     row += 1
                     bgr = row % 2
                     node = Node(columns[col],
-                         colorFalse=colorsFalse[bgr][bgc],
-                         colorTrue=colorsTrue[bgr][bgc]
+                         colorFalse=colorsBackground[bgr][bgc],
+                         colorTrue=colorsCharacter[bgr][bgc]
                          )
                     node.pack(fill=tk.X)
-                    self.sceneCharacters[scId][crId] = node
+                    self._characterNodes[scId][crId] = node
+
+        #--- Location columns.
+        self._locationNodes = {}
+        for chId in novel.chapters:
+            for scId in novel.chapters[chId].srtScenes:
+                self._locationNodes[scId] = {}
+        for lcId in novel.locations:
+            row = 0
+            bgr = row % 2
+            col += 1
+            bgc = col % 2
+            columns.append(tk.Frame(master))
+            columns[col].pack(side=tk.LEFT)
+            tk.Label(columns[col],
+                     text=novel.locations[lcId].title,
+                     bg=colorsBackground[bgr][bgc],
+                     justify=tk.LEFT,
+                     anchor=tk.W
+                     ).pack(fill=tk.X)
+            for chId in novel.chapters:
+                for scId in novel.chapters[chId].srtScenes:
+                    row += 1
+                    bgr = row % 2
+                    node = Node(columns[col],
+                         colorFalse=colorsBackground[bgr][bgc],
+                         colorTrue=colorsLocation[bgr][bgc]
+                         )
+                    node.pack(fill=tk.X)
+                    self._locationNodes[scId][lcId] = node
+
+        #--- Item columns.
+        self._itemNodes = {}
+        for chId in novel.chapters:
+            for scId in novel.chapters[chId].srtScenes:
+                self._itemNodes[scId] = {}
+        for itId in novel.items:
+            row = 0
+            bgr = row % 2
+            col += 1
+            bgc = col % 2
+            columns.append(tk.Frame(master))
+            columns[col].pack(side=tk.LEFT)
+            tk.Label(columns[col],
+                     text=novel.items[itId].title,
+                     bg=colorsBackground[bgr][bgc],
+                     justify=tk.LEFT,
+                     anchor=tk.W
+                     ).pack(fill=tk.X)
+            for chId in novel.chapters:
+                for scId in novel.chapters[chId].srtScenes:
+                    row += 1
+                    bgr = row % 2
+                    node = Node(columns[col],
+                         colorFalse=colorsBackground[bgr][bgc],
+                         colorTrue=colorsItem[bgr][bgc]
+                         )
+                    node.pack(fill=tk.X)
+                    self._itemNodes[scId][itId] = node
 
     def set_nodes(self):
-        for scId in self.sceneCharacters:
+        """Loop through all nodes, setting states."""
+        for scId in self._characterNodes:
             for crId in self._novel.characters:
                 try:
-                    self.sceneCharacters[scId][crId].state = (crId in self._novel.scenes[scId].characters)
+                    self._characterNodes[scId][crId].state = (crId in self._novel.scenes[scId].characters)
+                except TypeError:
+                    pass
+
+        for scId in self._locationNodes:
+            for lcId in self._novel.locations:
+                try:
+                    self._locationNodes[scId][lcId].state = (lcId in self._novel.scenes[scId].locations)
+                except TypeError:
+                    pass
+
+        for scId in self._itemNodes:
+            for itId in self._novel.items:
+                try:
+                    self._itemNodes[scId][itId].state = (itId in self._novel.scenes[scId].items)
                 except TypeError:
                     pass
 
     def get_nodes(self):
-        for scId in self.sceneCharacters:
+        """Loop through all nodes, modifying the scenes according to the states."""
+        for scId in self._characterNodes:
             self._novel.scenes[scId].characters = []
             for crId in self._novel.characters:
                 try:
-                    node = self.sceneCharacters[scId][crId]
+                    node = self._characterNodes[scId][crId]
                 except TypeError:
                     pass
                 else:
                     if node.state:
                         self._novel.scenes[scId].characters.append(crId)
 
+        for scId in self._locationNodes:
+            self._novel.scenes[scId].locations = []
+            for lcId in self._novel.locations:
+                try:
+                    node = self._locationNodes[scId][lcId]
+                except TypeError:
+                    pass
+                else:
+                    if node.state:
+                        self._novel.scenes[scId].locations.append(lcId)
 
-class Node(tk.Label):
-    """A matrix node, representing a boolean value."""
-    SIZE = 1
+        for scId in self._itemNodes:
+            self._novel.scenes[scId].items = []
+            for itId in self._novel.items:
+                try:
+                    node = self._itemNodes[scId][itId]
+                except TypeError:
+                    pass
+                else:
+                    if node.state:
+                        self._novel.scenes[scId].items.append(itId)
 
-    def __init__(self, master=None, colorFalse='white', colorTrue='black', cnf={}, **kw):
-        self.colorTrue = colorTrue
-        self.colorFalse = colorFalse
-        self._state = False
-        kw['width'] = self.SIZE
-        kw['height'] = self.SIZE
-        super().__init__(master, cnf, **kw)
-        self._set_color()
-        self.bind('<Button-1>', self._toggle_state)
-
-    @property
-    def state(self):
-        return self._state
-
-    @state.setter
-    def state(self, newState):
-        self._state = newState
-        self._set_color()
-
-    def _set_color(self):
-        if self._state:
-            self.config(background=self.colorTrue)
-        else:
-            self.config(background=self.colorFalse)
-
-    def _toggle_state(self, event=None):
-        self.state = not self._state